Remove v2 namespace for classifier
