+++ spec/lib/hackney/income/tenancy_classification/helpers/ma_agreement_helpers_spec.rb
@@ -1,8 +1,8 @@
 require 'rails_helper'
 
-describe Hackney::Income::TenancyClassification::V2::Helpers::MAAgreementHelpers do
+describe Hackney::Income::TenancyClassification::Helpers::MAAgreementHelpers do
   class DummyAgreementHelperClass
-    include Hackney::Income::TenancyClassification::V2::Helpers::HelpersProxy
+    include Hackney::Income::TenancyClassification::Helpers::HelpersProxy
 
     def initialize(case_priority, criteria, documents)
       @case_priority = case_priority
