+++ spec/support/shared_example_for_tenancy_classification_with_agreements_in_ma.rb
@@ -24,7 +24,7 @@ end
 # Alternatively, see any file that uses the Shared Example and see what they are supplying.
 #
 shared_examples 'TenancyClassificationWithAgreementsInMA' do |condition_matrix|
-  describe Hackney::Income::TenancyClassification::V2::Classifier do
+  describe Hackney::Income::TenancyClassification::Classifier do
     it_behaves_like 'TenancyClassification examples', condition_matrix
   end
 end
